Enabled travis CI support and fixed compiler warnings
Travis CI will build RBDL once with SimpleMath support and once with
Eigen3. The benchmark addon is included to validate performance on
Travis CI infrastructure. The builds use either GCC or clang to compile.
The pre_script performs configuration of the build via cmake defines.
The test script itself 1) compiles RBDL with "-Wall -Wpedantic -Wextra
-Werror", 2) runs the unit tests, and 3) runs the benchmark suite.

As part of the work, all warnings for RBDLi, benchmark and unittests were fixed.
Additionally the cmake infrastructure required small adaptions to build
everyting with c++11 support, i.e., rbdk, tests, benchmarks.

+++ tests/ModelTests.cc
@@ -61,7 +61,7 @@ TEST_FIXTURE(ModelFixture, TestAddBodyDimensions) {
   Joint joint ( SpatialVector (0., 0., 1., 0., 0., 0.));
 
   unsigned int body_id = 0;
-  body_id = model->AddBody(0, Xtrans(Vector3d(0., 0., 0.)), joint, body); 
+  body_id = model->AddBody(0, Xtrans(Vector3d(0., 0., 0.)), joint, body);
 
   CHECK_EQUAL (1u, body_id);
   CHECK_EQUAL (2u, model->lambda.size());
@@ -124,13 +124,13 @@ TEST_FIXTURE(ModelFixture, TestFloatingBodyDimensions) {
   CHECK_EQUAL (3u, model->mBodies.size());
 }
 
-/** \brief Tests whether the joint and body information stored in the Model are computed correctly 
+/** \brief Tests whether the joint and body information stored in the Model are computed correctly
 */
 TEST_FIXTURE(ModelFixture, TestAddBodySpatialValues) {
   Body body;
   Joint joint ( SpatialVector (0., 0., 1., 0., 0., 0.));
 
-  model->AddBody(0, Xtrans(Vector3d(0., 0., 0.)), joint, body); 
+  model->AddBody(0, Xtrans(Vector3d(0., 0., 0.)), joint, body);
 
   SpatialVector spatial_joint_axis(0., 0., 1., 0., 0., 0.);
   CHECK_EQUAL (spatial_joint_axis, joint.mJointAxes[0]);
@@ -142,7 +142,7 @@ TEST_FIXTURE(ModelFixture, TestAddBodyTestBodyName) {
   Body body;
   Joint joint ( SpatialVector (0., 0., 1., 0., 0., 0.));
 
-  model->AddBody(0, Xtrans(Vector3d(0., 0., 0.)), joint, body, "mybody"); 
+  model->AddBody(0, Xtrans(Vector3d(0., 0., 0.)), joint, body, "mybody");
 
   unsigned int body_id = model->GetBodyId("mybody");
 
@@ -200,7 +200,7 @@ TEST_FIXTURE(ModelFixture, TestjcalcSimple) {
 TEST_FIXTURE ( ModelFixture, TestTransformBaseToLocal ) {
   Body body;
 
-  unsigned int body_id = model->AddBody (0, SpatialTransform(), 
+  unsigned int body_id = model->AddBody (0, SpatialTransform(),
       Joint (
         SpatialVector (0., 0., 0., 1., 0., 0.),
         SpatialVector (0., 0., 0., 0., 1., 0.),
@@ -250,8 +250,8 @@ TEST ( Model2DoFJoint ) {
   Model model_std;
   model_std.gravity = Vector3d (0., -9.81, 0.);
 
-  model_std.AddBody(0, Xtrans(Vector3d(1., 0., 0.)), joint_rot_z, null_body); 
-  model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_x, body); 
+  model_std.AddBody(0, Xtrans(Vector3d(1., 0., 0.)), joint_rot_z, null_body);
+  model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_x, body);
 
   // using a model with a 2 DoF joint
   Joint joint_rot_zx (
@@ -293,13 +293,13 @@ TEST ( Model3DoFJoint ) {
 
   // in total we add two bodies to make sure that the transformations are
   // correct.
-  model_std.AddBody(0, Xtrans(Vector3d(1., 0., 0.)), joint_rot_z, null_body); 
-  model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_y, null_body); 
-  body_id = model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_x, body); 
+  model_std.AddBody(0, Xtrans(Vector3d(1., 0., 0.)), joint_rot_z, null_body);
+  model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_y, null_body);
+  body_id = model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_x, body);
 
-  model_std.AddBody(body_id, Xtrans(Vector3d(1., 0., 0.)), joint_rot_z, null_body); 
-  model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_y, null_body); 
-  body_id = model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_x, body); 
+  model_std.AddBody(body_id, Xtrans(Vector3d(1., 0., 0.)), joint_rot_z, null_body);
+  model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_y, null_body);
+  body_id = model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_x, body);
 
   // using a model with a 2 DoF joint
   Joint joint_rot_zyx (
@@ -353,9 +353,9 @@ TEST ( Model6DoFJoint ) {
       );
   body_id = model_std.AddBody (0, SpatialTransform(), joint_floating_base, body);
 
-  model_std.AddBody(body_id, Xtrans(Vector3d(1., 0., 0.)), joint_rot_z, null_body); 
-  model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_y, null_body); 
-  body_id = model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_x, body); 
+  model_std.AddBody(body_id, Xtrans(Vector3d(1., 0., 0.)), joint_rot_z, null_body);
+  model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_y, null_body);
+  body_id = model_std.AppendBody(Xtrans(Vector3d(0., 0., 0.)), joint_rot_x, body);
 
   // using a model with a 2 DoF joint
   Joint joint_rot_zyx (
@@ -403,7 +403,7 @@ TEST ( ModelFixedJointQueryBodyId ) {
   CHECK_EQUAL (fixed_body_id, model.GetBodyId("fixed_body"));
 }
 
-/* 
+/*
  * Makes sure that when appending a body to a fixed joint the parent of the
  * newly added parent is actually the moving body that the fixed body is
  * attached to.
@@ -614,8 +614,8 @@ TEST (TestAddFixedBodyToRoot) {
   // Add a mobile boby
   unsigned int movable_body = model.AppendBody (Xrotx (45 * M_PI / 180), JointTypeRevoluteX, body, "MovableBody");
 
-  CHECK_EQUAL (2, model.mBodies.size());
-  CHECK_EQUAL (2, model.mFixedBodies.size());
+  CHECK_EQUAL ((unsigned int) 2, model.mBodies.size());
+  CHECK_EQUAL ((unsigned int) 2, model.mFixedBodies.size());
 
   VectorNd q = VectorNd::Zero(model.q_size);
